+++ eva/src/js/componentsSettings.js
@@ -5,9 +5,9 @@ import {  mdiGraph, mdiGestureTap, mdiBrush, mdiChartLine,   mdiPoll, mdiViewGri
 export default {  
   tools: [
     {name: 'Таблица', img: mdiTableLarge, type: 'table'},
-    {name: 'Линейный график', img: mdiChartLine, type: 'lineChart'},
+    //{name: 'Линейный график', img: mdiChartLine, type: 'lineChart'},
     {name: 'Мультилинейный график', img: mdiChartMultiline, type: 'multiLine'},
-    {name: 'Cтолбцовая диаграмма', img: mdiPoll , type: 'barchart'},
+    //{name: 'Cтолбцовая диаграмма', img: mdiPoll , type: 'barchart'},
     {name: 'Выпадающий список', img: mdiFormatListBulleted, type: 'select'},
     {name: 'Выбор времени', img: mdiCalendarMonth, type: 'picker'},
     {name: 'Граф', img: mdiGraph , type: 'graph'},
@@ -27,10 +27,10 @@ export default {
       width: 480,
       height: 205
     },
-    lineChart: {
-      width: 700,
-      height: 400
-    },
+    // lineChart: {
+    //   width: 700,
+    //   height: 400
+    // },
     graph: {
       width: 1000,
       height: 600
@@ -59,10 +59,10 @@ export default {
       width: 400,
       height: 80
     },
-    barchart: {
-      width: 700,
-      height: 450
-    },
+    // barchart: {
+    //   width: 700,
+    //   height: 450
+    // },
     tile: {
       width: 470,
       height: 400
@@ -74,7 +74,7 @@ export default {
   },
   icons: {
     table: mdiTableLarge,
-    lineChart: mdiChartLine,
+    //lineChart: mdiChartLine,
     multiLine: mdiChartMultiline,
     graph: mdiGraph,
     guntt: mdiChartGantt,
@@ -83,14 +83,14 @@ export default {
     single: mdiNumeric,
     button: mdiGestureTap,
     textarea: mdiCardTextOutline,
-    barchart: mdiPoll,
+  //  barchart: mdiPoll,
     tile: mdiViewGrid,
     csvg: mdiBrush,
   }, 
   options: {
-    lineChart: ['visible','level','boxShadow','lastDot'],
-    multiLine: ['visible','level','boxShadow','united','lastDot','metrics'], 
-    barchart: ['visible','level','boxShadow'],
+    //lineChart: ['visible','level','boxShadow','lastDot'],
+    multiLine: ['visible','level','boxShadow','timeFormat','united','lastDot','metrics'], 
+    //barchart: ['visible','level','boxShadow'],
     table: ['visible','level','boxShadow','rowcolor','columncolor','cellcolor'],
     select: ['visible','level','boxShadow','multiple'],
     picker: ['visible','level'],
@@ -102,24 +102,24 @@ export default {
     tile: ['visible','level','widthTile','heightTile'],
     csvg: ['visible','level','tooltip']
   },
-  reporstElements: ['table','lineChart','multiLine','barchart','guntt','tile','csvg'],
+  reporstElements: ['table','multiLine','guntt','tile','csvg'],
   reports: {
     table: {
       tooltip: 'Таблица',
       icon: mdiTableLarge
     },
-    lineChart: {
-      tooltip: 'Линейный график',
-      icon: mdiChartLine
-    },
+    // lineChart: {
+    //   tooltip: 'Линейный график',
+    //   icon: mdiChartLine
+    // },
     multiLine: {
       tooltip: 'Мультилинейный график',
       icon: mdiChartMultiline
     },
-    barchart: {
-      tooltip: 'Столбцовая диаграмма',
-      icon: mdiPoll
-    },
+    // barchart: {
+    //   tooltip: 'Столбцовая диаграмма',
+    //   icon: mdiPoll
+    // },
     graph: {
       tooltip: 'Граф',
       icon: mdiGraph