+++ CHANGELOG.md
@@ -42,6 +42,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - check datasource in index db and if haven`t start search
 
+- add timeFormet in multiline and remove linechart and barchart
+
 ## [1.44.26] [newSetting]
 ### Adde